Reword and clarify file transfer protocol
Mostly reformulation and indentation. A few addenda are made to the protocol, making it more strict. Those changes are technically not backwards-compatible,
therefore the original text is left there with a **very strong** suggestion to do things differently. In practise, this should not break
anything since all known implementations already don't do everything that is allowed.