Server-protocol: Tell mailbox in "allocated"
The server has already opened the mailbox at this point so there is
a mailbox ID, but the client would still need to issue an `open` in
order to access it. Now the "allocated" response also contains the
mailbox ID, so clients may save a round trip if they want to.